cleanup and removing cruft
