+++ crack-core/src/commands/admin/broadcast_voice.rs
@@ -1,8 +1,5 @@
-use serenity::builder::CreateMessage;
-
-use crate::utils::get_current_voice_channel_id;
-use crate::Context;
-use crate::Error;
+use crate::{Context, ContextExt, Error};
+use serenity::all::CreateMessage;
 
 /// Broadcast a message to all guilds where the bot is currently in a voice channel.
 #[poise::command(prefix_command, owners_only, ephemeral)]
@@ -12,22 +9,17 @@ pub async fn broadcast_voice(
     #[description = "The message to broadcast"]
     message: String,
 ) -> Result<(), Error> {
-    let data = ctx.data();
-    let http = ctx.http();
-    let serenity_ctx = ctx.serenity_context().clone();
-    let guilds = data.guild_settings_map.read().await.clone();
+    let guilds = ctx.data().guild_settings_map.read().await.clone();
 
     for (guild_id, _settings) in guilds.iter() {
-        let message = message.clone();
-
-        let channel_id_opt = get_current_voice_channel_id(&serenity_ctx, *guild_id).await;
-
-        if let Some(channel_id) = channel_id_opt {
-            channel_id
-                .send_message(&http, CreateMessage::new().content(message.clone()))
-                .await
-                .unwrap();
-        }
+        match ctx.get_active_channel_id(*guild_id).await {
+            Some(channel_id) => {
+                let _ = channel_id
+                    .send_message(ctx.http(), CreateMessage::new().content(message.clone()))
+                    .await;
+            },
+            None => tracing::warn!("No active channel for guild_id: {}", guild_id),
+        };
     }
 
     Ok(())
